Refactor app: fix store imports
