+++ src/components/modals/Modal.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import classNames from 'classnames/bind';
 
-import { closeModal } from '../../slices/modalSlice';
-import selectShow, { selectModal } from '../../slices/modalSelectors';
+import { closeModal } from '../../store/slices/modalSlice';
+import selectShow, { selectModal } from '../../store/selectors/modalSelectors';
 
 import ModalForm from './ModalForm';
