+++ src/init.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 
-import store from './slices/index.js';
+import store from './store/index.js';
 
 import App from './App.jsx';
